Move RegularizationTypeParser
