+++ datadriven/src/sgpp/datadriven/configuration/RegularizationTypeParser.hpp
